requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
dnspython==2.6.1
pydantic==2.6.3
email-validator==2.1.1
